重构 session 为 u64 并移除墓碑机制
隔离服务 panic，防止初始化卡死与 worker 被用户代码打崩

修复 ghost name、drain 自唤醒死循环与任务唤醒 ABA

收工后拒绝创建服务并加强邮箱状态机不变量检查

补充 panic、session、任务代际与 drain 回归测试

同步更新 timer、net、http、tls、cluster 的 session 类型

+++ crates/rskynet-cluster/src/lib.rs
@@ -488,7 +488,7 @@ impl std::fmt::Display for ServiceKey {
 
 struct Pending {
     source: u32,
-    session: i32,
+    session: u64,
     node: NodeId,
     response_type: u32,
 }
